src/requirements_test.txt
-r requirements.txt
wget
pylint
pytest
pytest-aiohttp
pytest-cov


